requirements.txt
duckdb
pyarrow
pandas
python-dotenv
pyyaml
requests
streamlit
